plugin-connector: split connector-coordinator into a directory
connector-coordinator.ts had grown to 826 lines covering several distinct
concerns (OAuth flow, pending-connection localStorage snapshots, single- vs
multi-target cursor creation, and the coordinator's own capability wiring).
Turn it into a directory, mirroring plugin-space's app-graph-builder and
plugin-native-filesystem's state capability: connector-coordinator/<name>.ts
holds the default-exported capability module, index.ts is a thin
Capability.lazy pointer to it, and oauth.ts / pending-snapshot.ts /
create-single-cursor.ts / reconcile-cursors.ts hold the split-out pieces,
each still specific to this one capability rather than the plugin's util/.

capabilities/index.ts and capabilities/node.ts now re-export the directory's
own index (`export * from './connector-coordinator'`) instead of each
independently wrapping the module in their own Capability.lazy, avoiding a
double-wrap now that the target is a directory.

reconcile-cursors.test.ts moves alongside reconcile-cursors.ts.
