+++ packages/plugins/plugin-connector/src/capabilities/connector-coordinator/index.ts
@@ -1,0 +1,7 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+import { Capability } from '@dxos/app-framework';
+
+export const Coordinator = Capability.lazy('ConnectorCoordinator', () => import('./connector-coordinator'));
