+++ packages/plugins/plugin-connector/src/capabilities/index.ts
@@ -5,10 +5,11 @@
 import { Capability } from '@dxos/app-framework';
 import { OperationHandlerSet } from '@dxos/compute';
 
+export * from './connector-coordinator';
+
 export const AppGraphBuilder = Capability.lazy('AppGraphBuilder', () => import('./app-graph-builder'));
 export const BuiltinConnectors = Capability.lazy('BuiltinConnectors', () => import('./connectors'));
 export const CreateObject = Capability.lazy('CreateObject', () => import('./create-object'));
-export const Coordinator = Capability.lazy('ConnectorCoordinator', () => import('./connector-coordinator'));
 export const OAuthRedirect = Capability.lazy('OAuthRedirect', () => import('./oauth-redirect'));
 export const OperationHandler = Capability.lazy<OperationHandlerSet.OperationHandlerSet>(
   'OperationHandler',
