+++ packages/plugins/plugin-connector/src/capabilities/node.ts
@@ -5,9 +5,10 @@
 import { Capability } from '@dxos/app-framework';
 import { OperationHandlerSet } from '@dxos/compute';
 
+export * from './connector-coordinator';
+
 export const AppGraphBuilder = Capability.lazy('AppGraphBuilder', () => import('./app-graph-builder'));
 export const BuiltinConnectors = Capability.lazy('BuiltinConnectors', () => import('./connectors'));
-export const Coordinator = Capability.lazy('ConnectorCoordinator', () => import('./connector-coordinator'));
 export const CreateObject = Capability.lazy('CreateObject', () => import('./create-object'));
 export const OperationHandler = Capability.lazy<OperationHandlerSet.OperationHandlerSet>(
   'OperationHandler',
